fix syntax problem

+++ tools_docker/Debian_Uwsgi/common.sh
@@ -42,6 +42,9 @@ fi
 if [ ! -f /etc/tracim/uwsgi.ini ]; then
     cp /tracim/uwsgi.ini.sample /etc/tracim/uwsgi.ini
 fi
+if [ ! -L /tracim/uwsgi.ini ];then
+    ln -s /etc/tracim/uwsgi.ini /tracim/uwsgi.ini
+fi
 if [ ! -L /etc/uwsgi/apps-available/tracim.ini ]; then
     ln -s /etc/tracim/uwsgi.ini /etc/uwsgi/apps-available/tracim.ini
 fi
@@ -87,6 +90,7 @@ if [ ! -f /etc/systemd/system/tracim_uwsgi.service ]; then
     cp /tracim/tools_docker/Debian_Uwsgi/new_service.service.sample /etc/systemd/system/tracim_uwsgi.service
     sed -i "s|Description=|Description=tracim_uwsgi service|g" /etc/systemd/system/tracim_uwsgi.service
     sed -i "s|ExecStart=|ExecStart=/usr/bin/uwsgi --ini /tracim/uwsgi.ini --http-socket :8080 --plugin python3 --uid www-data --gid www-data|g" /etc/systemd/system/tracim_uwsgi.service
+    chmod +x /etc/systemd/system/tracim_uwsgi.service
 fi
 
 # Webdav
@@ -94,12 +98,17 @@ if [ ! -f /etc/systemd/system/tracim_webdav.service ]; then
     cp /tracim/tools_docker/Debian_Uwsgi/new_service.service.sample /etc/systemd/system/tracim_webdav.service
     sed -i "s|Description=|Description=tracim_webdav service|g" /etc/systemd/system/tracim_webdav.service
     sed -i "s|ExecStart=|ExecStart=/usr/bin/uwsgi --ini /tracim/webdav.ini --http-socket :3030|g" /etc/systemd/system/tracim_webdav.service
+    chmod +x /etc/systemd/system/tracim_webdav.service
 fi
 # Webdav config
 if [ "$WEBDAV" = "start" ]; then
-    cp /tracim/uwsgi.ini.sample /etc/tracim/webdav.ini
-    ln -s /etc/tracim/webdav.ini /tracim/webdav.ini
-    sed -i "s|module = wsgi.web:application|module = wsgi.webdav:application|g" /etc/tracim/webdav.ini
+    if [ ! -f /etc/tracim/webdav.ini ];then
+        cp /tracim/uwsgi.ini.sample /etc/tracim/webdav.ini
+    fi
+    if [ ! -L /tracim/webdav.ini ];then
+        ln -s /etc/tracim/webdav.ini /tracim/webdav.ini
+        sed -i "s|module = wsgi.web:application|module = wsgi.webdav:application|g" /etc/tracim/webdav.ini
+    fi
 fi
 if [ "$WEBDAV" = "start" ]; then
     sed -i "s|#<Directory "/">|<Directory "/">|g" /etc/tracim/apache2.conf
